pyproject.toml
[project]
name = "tg-bot-icecream-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "gergpav",email = "gergpav2018@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-telegram-bot[job-queue] >=20.0,<21.0",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "types-psycopg2 (>=2.9.21.20250516,<3.0.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "ptbcontrib[roles] @ git+https://github.com/python-telegram-bot/ptbcontrib.git@main",
    "uvicorn (>=0.35.0,<0.36.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.16.1"
asyncpg-stubs = "^0.30.1"

